[uart] implement otPlatUartFlush() (#51)
This commit is a proposition of a problem reported in
openthread/openthread#6623. Current CLI implementation (UART variant)
uses otPlatUartFlush() for corner case and crashes when it is not
implemented. While it would be better to handle asynchronous
transmissions in CLI I think that there is no strong reason to not
implement this function.