[spi] fixing an issue with spi transaction being too short
Issue #7224 was reported that aTransactionLength parameter of SPI
otPlatSpiSlaveTransactionCompleteCallback callback is used incorrectly
by passing only received buffer size instead of real transaction
length.
The driver does not provide functionality to get this value, however in
this case providing maximum of the received and transmitted lengths
should do the trick.

Signed-off-by: Marek Porwisz <marek.porwisz@nordicsemi.no>