[crypto] add crypto platform API (#369)
The radio driver calls the `nrf_802154_tx_started` when the SFD is
sent. The radio driver must complete the encryption of the first byte
of the payload before it is sent in the function
`nrf_802154_tx_started`.  The chip nrf52811 is using the software AES
ECB encryption, the encryption speed is slow. Sometimes, we observed
that the first received payload byte is not encrypted.

This commit uses the ECB hardware to implement the crypto platform API
to improve the encryption speed.