[usb] fix interface numbering with DFU support disabled (#580)
Using ot-nrf528xx without DFU support causes the Linux kernel to issue
the following warning:

  usb 2-3.1.3: new full-speed USB device number 105 using xhci_hcd
  usb 2-3.1.3: config 1 has an invalid interface number: 2 but max is 1
  usb 2-3.1.3: config 1 has no interface number 0
  usb 2-3.1.3: New USB device found, idVendor=1915, idProduct=cafe, bcdDevice= 1.00
  usb 2-3.1.3: New USB device strings: Mfr=1, Product=2, SerialNumber=3
  usb 2-3.1.3: Product: nRF528xx OpenThread Device
  usb 2-3.1.3: Manufacturer: Nordic Semiconductor

According to the USB specification, interface numbers should start at 0
and cannot have gaps. The DFU interface has number 0, while the CDC
interfaces are 1 and 2. Therefore when DFU is disabled, the first interface
does not have the number 0.

The fix is to shift the CDC interfaces to interface numbers 0 and 1, as
they are always present, and move the DFU interface to number 2, as it
might be present or not depending on the ot-nrf528xx enabled features.