[nrf528xx] clear the flag psdu_being_received in CRCOK handler
The flag psdu_being_received is set to True when the radio driver receives
the FCF field of the data frame, and the flag will be cleared after the
ACK is sent out if the ACK is required. Sometimes, the ACK is failed to be
sent out, which causes the flag to be set to True forever. Then the radio
driver can't send out any frames.

The sdk-nrfxlib (https://github.com/nrfconnect/sdk-nrfxlib/blob/main/nrf_802154/driver/src/nrf_802154_trx.c#L2492)
clears the flag `psdu_being_received` in the CRCOK handler. This commit clears
the flag `psdu_being_received` when the irq_crcok_state_rx() is handled.

Ref: https://nrfsupport.atlassian.net/browse/NRF11-73