[radio] correct the `mInfo.mRxInfo.mTimestamp` value of recevied frames (#966)
The `mInfo.mRxInfo.mTimestamp` represents the end time of the SFD of the
received frame. But the radio driver sets the `mInfo.mRxInfo.mTimestamp`
value to the first symbol time of the received frame.

This commit adds a method `nrf_802154_timestamp_end_to_phr_convert()`
to calculate the end time of the SFD of the received frame. The method
comes from the `sdk-nrfxlib`.